Modification de 'price' en float

+++ src/main/java/com/carrot/carrotshop/shop/Sell.java
@@ -36,7 +36,7 @@ public class Sell extends Shop {
 	@Setting
 	private Location<World> sellerChest;
 	@Setting
-	private int price;
+	private float price;
 
 	public Sell() {
 	}
